Remove memory leaks; local copy of thermal img
Removed memory leaks, rewrote image reading from camera to create
a local copy of the temperature image to have complete control over
memory management.

The WIC SDK only has functions to free the last image of the camera
 buffer, but keeps track of buffers still in use
(although it doesn't allow access to them) and tries to free them
when the camera is disconnected.

Copying the thermal image allows full control over when the image
 memory is freed. As the image is small (around 168 kiB for a
256*336 camera), it is also negligibly fast, especially  compared
 to the image processing applications.
The memory in the camera buffer is freed immediately after copy.

Closes #14