Update WIC SDK to recently released version 1.2.1
This seems to fix the problem with camera calibration :-)

We have also refactored few .nix files, mainly to take advantage of
autoPatchelfHook. With this, we could get rid of setting
the LD_LIBRARY_PATH variable (in the run script and the systemd
service unit), but we keep it there, because it might be needed for
people brave enough to run build the software without Nix.