Generate images for web and their descriptions dynamically
Now, when we add or remove an image, it's not necessary to update the
code at multiple places (C++, html, javascript). It's sufficient to
add the image to webimgs list and the rest is done automatically.

In the future, this will also allow downloading of original
images (with float-valued pixels), rather than RGB visualisations.
This will be needed for more automated processing of captured data.