Copy reference points to current image when tracking is temporarily off
This helps in situation, when the board is switched off and the
program is restarted. Previously, no POIs and heat sources were
measured after restart, because they get initialized only after
tracking is active, which it never is when the board is off due to low
temperature difference. Now, we just copy the reference points to the
current image. Position of the points might be slightly off, but it's
better than nothing. Once the board is switched on and the tracking
gets activated, the positions will be corrected.