Exit on image freeze
I hoped that "image freezing" that we experienced with the old SDK is
gone. Unfortunately, it happens even with WIC SDK 2.0.

Previously, with SDK 1.2.1, we detected the situation prior to "image
freezing" by receiving the same buffer pointer from two subsequent
calls to camera->retrieveBuffer(). Then it was sufficient to stop and
start acquisition to recover from this situation. However, sometimes,
the retrieveBuffer blocked indefinitely without signalizing the
situation with same pointers in advance. In that case systemd watchdog
saved us by restarting the program.

With the new SDK, we cannot detect two subsequent pointers being the
same, because we don't get access to them. However, the buffers
returned by the new SDK start to have zero length. In this case, we
just exit to let systemd restart us.